requirements/base.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
gunicorn>=22.0,<24.0
redis>=5.0,<7.0
django-unfold>=0.40.0,<1.0
django-apscheduler>=0.7.0,<1.0
